backend/requirements.txt
bcrypt==5.0.0
bleach==6.2.0
blinker==1.9.0
cachelib==0.13.0
cairocffi==1.4.0
click==8.3.0
cryptography==43.0.1
docusign-esign==3.24.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
msgspec==0.19.0
packaging==25.0
psycopg2-binary==2.9.9
pydyf==0.9.0
python-dotenv==1.1.1
pyjwt==2.10.0
redis==6.4.0
requests
SQLAlchemy==2.0.40
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.3
WeasyPrint==61.2